migrate tests to v2
